ChatServer.java specific feedback
Some comments of interest - you don't have to use so many intermediate variables (although it's possible) and, at the least, you should not be storing every intermediate variable in the class. At most InputStream and BufferedReader could be class level, and I'd argue that InputStream (since it's closed and handled by BufferedReader) should not be stored there.